Create sample/utils directory and refactor
