+++ sample/model/n_hot_from_scratch.py
