+++ sample/solver/_solver_helper.py
